refactor: split command projection from cli grammar
