+++ src/commands/cli-grammar.ts
@@ -1,8 +1,8 @@
+export { readInputFromCli } from './cli-grammar/registry.ts';
 export {
   prepareBatchStep,
   prepareDaemonCommandRequest,
-  readInputFromCli,
   batchCommandNames,
   type BatchCommandName,
   type DaemonCommandName,
-} from './cli-grammar/registry.ts';
+} from './command-projection.ts';
